Project Finish
For now, all the important parts of this project is finished. I will probably go in and comment later, as well as do some optimization.
This commit splits the 3 body simulation into two files (while preserving the original). This is done so that the visual representation of the simulation can be sped up, as the simulation uses small timesteps.
Be warned, the log files take up a fair amount of space (100 MB)